Rough complete

+++ frontend/src/components/account/index.vue
@@ -1,8 +1,8 @@
 <template>
 	<div class="hello">
-		<b-container>
+		<b-container fluid>
 			<b-row class="my-4">
-				<b-col cols="4">
+				<b-col cols="8" sm="6">
 					<b-card :title="privates.CustomerID" 
 							:sub-title="`${privates.CustomerName} (${privates.Age}, ${privates.Gender})`">
 						<p class="mb-1 card-text">
@@ -18,7 +18,8 @@
 				</b-col>
 			</b-row>
 			<b-row class="mt-2">
-				<b-table v-if="reservations.length > 0" hover :items="reservations" :fields="fields"/>
+				<b-table v-if="reservations.length > 0" hover :items="reservations" :fields="fields"
+						@row-clicked="cancelReservation"/>
 				<h2 class="mx-3 my-3" v-else>You have no reservations.</h2>
 			</b-row>
 		</b-container>
@@ -40,6 +41,11 @@ export default {
 				{
 					key: 'ReservationID',
 					label: 'ID',
+					sortable: true
+				},
+				{
+					key: 'MovieTitle',
+					label: 'Movie',
 					sortable: false
 				},
 				{
@@ -61,6 +67,16 @@ export default {
 					key: 'ReservedDate',
 					label: 'Reserved At',
 					sortable: true
+				},
+				{
+					key: 'CinemaName',
+					label: 'Cinema',
+					sortable: true
+				},
+				{
+					key: 'RoomName',
+					label: 'Room',
+					sortable: false
 				},
 				{
 					key: 'Seats',
@@ -111,13 +127,18 @@ export default {
 				}
             })
 		},
-		cancelReservation(resID) {
+		cancelReservation(item) {
+			var resID = item.ReservationID;
+			var answer = confirm(`Cancel reservation for ${resID}?`);
+			if (!answer) {
+				return;
+			}
             this.$http.delete(`api/reservate/${resID}`, {
                 headers: {
                     'Authorization': `Bearer ${sessionStorage['token']}`
                 }
             }).then((res) => {
-				
+				this.reservations = this.reservations.filter(x => x.ReservationID != resID);
             }).catch(err => {
                 var msg = '';
                 if (err.response) { // Response Error
